I finished flex module, which means I finished the css foundations!
This was a bit difficult at first, some things I was not sure about,
but in the end I got it. It was mostly figuring out how default
properties worked and how this would change the items structure
and spacing. In the end I got mostly everything by myself, just
with some work. I also need to look for easier solutions to things.
This was fun!